Re-organized and renamed folders, files and tests. Added one extra check to search.spec.ts first test
