+++ pages/DeleteAccountPage.ts
